refactor: Overlay 컴포넌트 Common 으로 이동
