+++ components/Common/Overlay.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+import { ReactElement } from 'react';
 import styled from '@emotion/styled';
 
 type OverlayPropType = {
@@ -17,3 +19,14 @@ export const OverlayContainer = styled.div<OverlayPropType>`
   z-index: 2;
   cursor: pointer;
 `;
+
+interface OverlayProps {
+  onClose: (e: React.MouseEvent) => void;
+  height: number;
+}
+
+const Overlay = ({ onClose, height }: OverlayProps): ReactElement => {
+  return <OverlayContainer height={height} onClick={onClose} />;
+};
+
+export default Overlay;
